some assets imported
Hey guys a lot of the assets we liked did not work importing into unity. realised that they all need to be in fbx format and even if they are the chances of them importing correctly are low. so ive imported only the ones that i could find that work